fix: handle currently available non-completed media (#4115)
* fix: check media status has changed before modifying request

fix: refactor: code cleanup

* fix: manually load database entity seasons
